Solution to Check MSB of a Number

+++ Check MSB of a Number@@67882a731b4392d8fc1577e9/code.c
@@ -5,8 +5,8 @@
 // }
 
 int main() {
-    int a ;
-    scanf("%d" ,&a);
+    long a ;
+    scanf("%ld" ,&a);
     int msb = num & (1 <<  (31));
     if (msb == 1)
     printf("Set");
